libinput: add gesture support in libinput handlers
