+++ include/taiwins/input_device.h
@@ -207,6 +207,20 @@ struct tw_event_pointer_axis {
 	int32_t delta_discrete;
 };
 
+struct tw_event_pointer_gesture {
+	struct tw_input_device *dev;
+	uint32_t time, fingers;
+
+	enum {
+		TW_POINTER_GESTURE_BEGIN = 0,
+		TW_POINTER_GESTURE_UPDATE,
+		TW_POINTER_GESTURE_END,
+	} state;
+
+	double dx, dy, rotation, scale;
+	bool cancelled;
+};
+
 /** touch events */
 struct tw_event_touch_down {
 	struct tw_input_device *dev;
