+++ libtaiwins/backend/libinput/meson.build
@@ -1,5 +1,6 @@
 taiwins_lib_src += files(
   'device.c',
+  'input.c',
 )
 
 taiwins_lib_dep += [
